Show loader for all cases where video doesn't autoplay
Displays the loader gif on Android as well as on iOS during the brief
interval where the video is visible but not yet autoplaying. It should
only be a fraction of a second, but showing activity seems better than
nothing.

Also apply some basic css to the video itself to match the behavior that
would be expected of an image (`height: auto`, so that it collapses to a
proportional scale and doesn't take up as much space as the `height`
attribute set on it.)