Get rid of loader gif, use css animation
That gif I chose was ugly as hell. This is a really basic animation, but
maybe someone with design chops can improve on it. It only has to be
good enough to show that something is happening, in case someone lands
on a page with a lot of gifs, and they don't start playing right away.