Abstract fallbackGif to its own function
The same functionality was being used in two separate places, to handle
errors and to handle iOS devices. Moved it up into a hidden helper
function inside VideoHandler, so that it can be called in both places.

Also moved the checkInView and startPlaying function up there as well,
as it didn't make sense for them to be inside the handleMobile function.